Implement swiping (for mobile) and dragging (for mouse users).
Grr, `touchend` events fire on the element where the touch STARTED!
This means we have to use `document.elementFromPoint` to figure out
which element was SUPPOSED to get the event, and then ask it what
its position is. The element that we get might be a `Tile` or it
might be a `grid` div.

https://stackoverflow.com/questions/11523496/find-element-finger-is-on-during-a-touchend-event
https://stackoverflow.com/questions/19179817/how-to-get-target-element-on-touch-end-event-handler