AI: Modified form of the position-dependent blueness boost from Suetsugu's paper.
This is table 7 in section 2.2, modified ad-hoc (with no particular math behind it)
to deal with the fact that we're playing PvP and not Geister. We add this boost
to the estimated blueness of each piece when we're thinking about capturing it.

Based on "Developping geister algorithms without machine learning"
(Koki Suetsugu, Yusuke Orita; 2018; full text in Japanese).
https://ipsj.ixsq.nii.ac.jp/ej/?action=repository_action_common_download&item_id=186127&item_no=1&attribute_id=1&file_no=1