AI: Abandon goalkeeping once 4 of my pieces are captured.
Otherwise it's too easy to end up in a situation where two pieces
are pinned in the goalkeeper positions, one piece is pinned two steps
from the goal (but it's red so it refuses to advance), and the fourth
piece twiddles back and forth in the back rank (because it's stuck
behind a goalkeeper).

However, it's still possible to get stuck behind a red piece that
refuses to advance — or, generally, at the tail end of a traffic jam
of blues lined up behind a red that refuses to advance.