AI: Use "patterns" in moveWins().
This code is much closer to what Suetsugu and Okita did in their paper.
We tabulate a set of "patterns" and match each pattern against the current
game state. If we find that the game state matches one of our patterns,
then we have an actually winning move.

Well, we might still lose the game by capturing too many red; and we might
lose the game by losing a horse-race with one of the human's blue pieces
(sadly we have no checks for "is the human racing a piece to the finish").
But except for THOSE possibilities, a pattern-match guarantees a winning move.