AI: Look one move ahead to avoid silly losing moves in the endgame.
With the old code, you could frequently beat the AI just by camping
out next to the "goalkeeper" square and attritting the enemy forces
until the AI enters "endgame mode" and randomly moves the goalkeeper
away; then you just walk in for the win.

This new logic takes the "winningPatterns" we've already encoded and
flips them to apply to the human player. We discard any move that
would lead to a position where the human player has a winning move.
(For this purpose, we assume that every human piece is blue.
We can do better, but this is a great start.)

The AI should no longer lose by "silly move in the endgame."